requirements.txt
fastapi[standard]
uvicorn
sqlmodel
pyjwt 
passlib[bcrypt]
openai-agents